refactor(stack): Propsの型定義を切り出し
