+++ packages/wiz-ui-react/src/components/base/stack/components/h-stack.tsx
@@ -1,6 +1,7 @@
 import { memo } from "react";
 
-import { Props as StackProps, WizStack } from "./stack";
+import { WizStack } from "./stack";
+import { StackProps } from "./types";
 
 type Props = Omit<StackProps, "direction">;
