+++ packages/wiz-ui-react/src/components/base/stack/components/types.ts
@@ -1,0 +1,32 @@
+import { SpacingKeys } from "@wizleap-inc/wiz-ui-constants";
+import { ReactNode } from "react";
+
+export type StackProps = {
+  direction?: "horizontal" | "vertical";
+  align?: "start" | "center" | "end" | "stretch";
+  justify?: "start" | "center" | "end" | "between" | "around" | "evenly";
+  wrap?: boolean;
+  gap?: SpacingKeys;
+  gx?: SpacingKeys;
+  gy?: SpacingKeys;
+  p?: SpacingKeys;
+  pt?: SpacingKeys;
+  pr?: SpacingKeys;
+  pb?: SpacingKeys;
+  pl?: SpacingKeys;
+  px?: SpacingKeys;
+  py?: SpacingKeys;
+  m?: SpacingKeys;
+  mt?: SpacingKeys;
+  mr?: SpacingKeys;
+  mb?: SpacingKeys;
+  ml?: SpacingKeys;
+  mx?: SpacingKeys;
+  my?: SpacingKeys;
+  width?: string;
+  height?: string;
+  overflow?: string;
+  reverse?: string;
+  position?: "relative" | "absolute" | "fixed" | "sticky" | "static";
+  children?: ReactNode;
+};
